Add button to create list on blank page
